+++ src/meeple/util/output_util.py
@@ -1,4 +1,5 @@
 import numbers
+from enum import Enum
 
 from rich import box
 from rich.console import Console
@@ -6,31 +7,66 @@ from rich.table import Table
 
 from meeple.util.api_util import BOARDGAME_TYPE, EXPANSION_TYPE
 
-NA = "[bright_black]NA[/bright_black]"
+NA_VALUE = "[bright_black]NA[/bright_black]"
+
+SORT_ASC_SYMBOL = "[blue]˄[/blue]"
+SORT_DESC_SYMBOL = "[blue]˅[/blue]"
+
+
+class ItemHeader(Enum):
+    COUNT = ("#", "count")
+    ID = ("ID", "id")
+    NAME = ("Name", "name")
+    TYPE = ("Type", "type")
+    COLLECTION = ("Collection(s)", "collection")
+    YEAR = ("Year", "year")
+    RANK = ("Rank", "rank")
+    RATING = ("Rating", "rating")
+    WEIGHT = ("Weight", "weight")
+    PLAYERS = ("Players", "players")
+    TIME = ("Play Time", "time")
+
+
+class CollectionHeader(Enum):
+    NAME = ("Name", "name")
+    BOARDGAMES = ("Boardgames", "boardgames")
+    EXPANSIONS = ("Expansions", "expansions")
+    UPDATED = ("Last Updated", "updated")
+
+
+def fmt_headers(headers: ItemHeader, sort_key: str, sort_direction: str):
+    header_strs = []
+    for header in headers:
+        if sort_key and header.value[1] == sort_key:
+            header_strs.append(f"{header.value[0]} {sort_direction}")
+            continue
+        header_strs.append(header.value[0])
+
+    return header_strs
 
 
 def fmt_players(minplayers: str, maxplayers: str) -> str:
     if int(minplayers) == int(maxplayers) == 0:
-        return NA
+        return NA_VALUE
     return f"{minplayers}-{maxplayers}"
 
 
 def fmt_playtime(playtime: str) -> str:
     if int(playtime) == 0:
-        return NA
+        return NA_VALUE
     return f"~{playtime} Min"
 
 
 def fmt_avg_rank(rank: str) -> str:
     if not isinstance(rank, numbers.Number) or int(rank) == 0:
-        return NA
+        return NA_VALUE
     rank_str = f"{rank:.2f}"
     return rank_str
 
 
 def fmt_rank(rank: str) -> str:
     if rank == "NA" or not rank.isdigit():
-        return NA
+        return NA_VALUE
     return rank
 
 
@@ -43,7 +79,7 @@ def fmt_rating(rating: float) -> str:
     if rating > 6:
         return f"[magenta]{rating_str}[/magenta]"
     if rating == 0:
-        return NA
+        return NA_VALUE
     return f"[red]{rating_str}[/red]"
 
 
@@ -52,7 +88,7 @@ def fmt_type(item_type: str) -> str:
         return "Board Game"
     if item_type == EXPANSION_TYPE:
         return "Expansion"
-    return NA
+    return NA_VALUE
 
 
 def fmt_weight(weight: float) -> str:
@@ -64,13 +100,13 @@ def fmt_weight(weight: float) -> str:
     if weight >= 2:
         return f"[bright_yellow]{weight_str}[/bright_yellow]"
     if weight == 0:
-        return NA
+        return NA_VALUE
     return f"[green]{weight_str}[/green]"
 
 
 def fmt_year(year: str) -> str:
     if int(year) == 0:
-        return NA
+        return NA_VALUE
     return year